// projects/ngx-epic-fhir-client/src/lib/model/bundleEndpointR4.ts
/**
 * Epic on Fhir
 * All of the components of the Change Healthcare Interoperability Apis
 *
 * OpenAPI spec version: 1.0.0
 * Contact: bmain@lumeris.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BundleEntryEndpointR4 } from './bundleEntryEndpointR4';

/**
 * Contains a collection of resources.
 */
export interface BundleEndpointR4 {
  /**
   * Entry in the bundle.
   */
  entry?: Array<BundleEntryEndpointR4>;
}