// projects/ngx-epic-fhir-client/src/lib/model/endpointR4.ts
/**
 * Epic on Fhir
 * All of the components of the Change Healthcare Interoperability Apis
 *
 * OpenAPI spec version: 1.0.0
 * Contact: bmain@lumeris.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CodeableConcept } from './codeableConcept';

/**
 * A Response Type
 */
export interface EndpointR4 { 
    /**
     * <p>The provider address. This is a direct address, not a general address.</p>
     */
    address?: string;
    /**
     * <p>This is always \"direct-project\".</p>
     */
    connectionType?: any;
    /**
     * <p>A name by which you can identify this endpoint.</p>
     */
    name: string;
    /**
     * <p>This element returns \"text/xml\", \"text/plain\", and \"application/pdf\".</p>
     */
    payloadMimeType: string;
    payloadType?: CodeableConcept;
    /**
     * <p>This is always \"active\".</p>
     */
    status: string;
}